fix: clean up urlscan sdk action structure
