+++ src/app.py
@@ -11,13 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import ipaddress
-import uuid
-
-from pydantic import field_validator
 from soar_sdk.app import App
 from soar_sdk.asset import BaseAsset, AssetField
-from soar_sdk.params import Param, Params
 
 from .actions.connectivity import run_test_connectivity
 from .actions.detonate import run_detonate_url
@@ -34,6 +29,13 @@ from .outputs import (
     ScreenshotActionOutput,
     ScreenshotSummary,
     UrlscanReportOutput,
+)
+from .params import (
+    DetonateUrlParams,
+    GetReportParams,
+    GetScreenshotParams,
+    LookupDomainParams,
+    LookupIpParams,
 )
 from .views import (
     render_detonate_url,
@@ -68,28 +70,6 @@ app = App(
 app.test_connectivity()(run_test_connectivity)
 
 
-class GetReportParams(Params):
-    id: str = Param(
-        description="Detonation ID for the desired report",
-        primary=True,
-        cef_types=["urlscan submission id"],
-    )
-
-    @field_validator("id")
-    @classmethod
-    def validate_id(cls, value: str) -> str:
-        candidate = value.strip()
-        try:
-            parsed_uuid = uuid.UUID(candidate)
-        except (TypeError, ValueError) as exc:
-            raise ValueError("Please provide a valid detonation ID") from exc
-
-        if str(parsed_uuid) != candidate.lower():
-            raise ValueError("Please provide a valid detonation ID")
-
-        return candidate
-
-
 app.register_action(
     run_get_report,
     identifier="get_report",
@@ -101,37 +81,6 @@ app.register_action(
     view_handler=render_get_report,
     view_template="get_report.html",
 )
-
-
-class LookupDomainParams(Params):
-    domain: str = Param(
-        description="Domain to lookup", primary=True, cef_types=["domain"]
-    )
-
-    @field_validator("domain")
-    @classmethod
-    def validate_domain(cls, value: str) -> str:
-        candidate = value.strip()
-        if not candidate:
-            raise ValueError("Please provide a valid domain")
-
-        if len(candidate) > 253:
-            raise ValueError("Please provide a valid domain")
-
-        if candidate.lower().startswith(("http://", "https://")):
-            raise ValueError("Please provide a valid domain")
-
-        try:
-            ipaddress.ip_address(candidate)
-        except ValueError:
-            pass
-        else:
-            raise ValueError("Please provide a valid domain")
-
-        if not any(char.isalnum() for char in candidate):
-            raise ValueError("Please provide a valid domain")
-
-        return candidate
 
 
 app.register_action(
@@ -146,24 +95,6 @@ app.register_action(
 )
 
 
-class LookupIpParams(Params):
-    ip: str = Param(description="IP to lookup", primary=True, cef_types=["ip", "ipv6"])
-
-    @field_validator("ip")
-    @classmethod
-    def validate_ip(cls, value: str) -> str:
-        candidate = value.strip() if isinstance(value, str) else str(value).strip()
-        if not candidate:
-            raise ValueError("Please provide a valid IPv4 or IPv6 address")
-
-        try:
-            ipaddress.ip_address(candidate)
-        except ValueError as exc:
-            raise ValueError("Please provide a valid IPv4 or IPv6 address") from exc
-
-        return candidate
-
-
 app.register_action(
     run_hunt_ip,
     name="lookup ip",
@@ -174,23 +105,6 @@ app.register_action(
     output_class=LookupActionOutput,
     summary_type=LookupSummary,
 )
-
-
-class DetonateUrlParams(Params):
-    url: str = Param(
-        description="URL to detonate", primary=True, cef_types=["url", "domain"]
-    )
-    tags: str | None = Param(
-        description="Comma-separated list of tags to annotate this scan. Limited to 10 tags. Tags with lengths longer than 29 will be omitted"
-    )
-    private: bool | None = Param(description="Run a private scan", default=False)
-    custom_agent: str | None = Param(description="Override User-Agent for this scan")
-    get_result: bool | None = Param(
-        description="Get scan result in same call", default=True
-    )
-    addto_vault: bool | None = Param(
-        description="Add url screenshot to vault", default=False
-    )
 
 
 app.register_action(
@@ -206,15 +120,6 @@ app.register_action(
     view_handler=render_detonate_url,
     view_template="detonate_url.html",
 )
-
-
-class GetScreenshotParams(Params):
-    report_id: str = Param(
-        description="UUID of report", primary=True, cef_types=["urlscan submission id"]
-    )
-    container_id: int | None = Param(
-        description="Event to add file to, will default to current container id"
-    )
 
 
 app.register_action(
